:sparkles: Set negative Ids to entities (#352)
* :dragon: Use longs for ids

* :sparkles: Add IdProvider

* :construction: Ensure new entities have their id set

* :green_heart:

* :white_check_mark: Test IdProvider in the Start method

+++ Toggl.Foundation/DataSources/Interfaces/ITimeEntriesSource.cs
@@ -13,6 +13,6 @@ namespace Toggl.Foundation.DataSources
 
         IObservable<ITimeEntry> Start(DateTimeOffset startTime, string description, bool billable);
 
-        IObservable<Unit> Delete(int id);
+        IObservable<Unit> Delete(long id);
     }
 }
