+++ Toggl.Foundation/Models/FoundationModels.cs
@@ -7,9 +7,9 @@ namespace Toggl.Foundation.Models
 {
     internal partial class Client : IDatabaseClient
     {
-        public int Id { get; }
+        public long Id { get; }
 
-        public int WorkspaceId { get; }
+        public long WorkspaceId { get; }
 
         public string Name { get; }
 
@@ -22,11 +22,11 @@ namespace Toggl.Foundation.Models
 
     internal partial class Project : IDatabaseProject
     {
-        public int Id { get; }
+        public long Id { get; }
 
-        public int WorkspaceId { get; }
+        public long WorkspaceId { get; }
 
-        public int? ClientId { get; }
+        public long? ClientId { get; }
 
         public string Name { get; }
 
@@ -59,9 +59,9 @@ namespace Toggl.Foundation.Models
 
     internal partial class Tag : IDatabaseTag
     {
-        public int Id { get; }
+        public long Id { get; }
 
-        public int WorkspaceId { get; }
+        public long WorkspaceId { get; }
 
         public string Name { get; }
 
@@ -72,15 +72,15 @@ namespace Toggl.Foundation.Models
 
     internal partial class Task : IDatabaseTask
     {
-        public int Id { get; }
+        public long Id { get; }
 
         public string Name { get; }
 
-        public int ProjectId { get; }
+        public long ProjectId { get; }
 
-        public int WorkspaceId { get; }
+        public long WorkspaceId { get; }
 
-        public int? UserId { get; }
+        public long? UserId { get; }
 
         public int EstimatedSeconds { get; }
 
@@ -95,13 +95,13 @@ namespace Toggl.Foundation.Models
 
     internal partial class TimeEntry : IDatabaseTimeEntry
     {
-        public int Id { get; }
+        public long Id { get; }
 
-        public int WorkspaceId { get; }
+        public long WorkspaceId { get; }
 
-        public int? ProjectId { get; }
+        public long? ProjectId { get; }
 
-        public int? TaskId { get; }
+        public long? TaskId { get; }
 
         public bool Billable { get; }
 
@@ -121,7 +121,7 @@ namespace Toggl.Foundation.Models
 
         public DateTimeOffset? ServerDeletedAt { get; }
 
-        public int UserId { get; }
+        public long UserId { get; }
 
         public string CreatedWith { get; }
 
@@ -132,11 +132,11 @@ namespace Toggl.Foundation.Models
 
     internal partial class User : IDatabaseUser
     {
-        public int Id { get; }
+        public long Id { get; }
 
         public string ApiToken { get; }
 
-        public int DefaultWorkspaceId { get; }
+        public long DefaultWorkspaceId { get; }
 
         public string Email { get; }
 
@@ -173,7 +173,7 @@ namespace Toggl.Foundation.Models
 
     internal partial class Workspace : IDatabaseWorkspace
     {
-        public int Id { get; }
+        public long Id { get; }
 
         public string Name { get; }
