+++ Toggl.Multivac/Models/ITimeEntry.cs
@@ -5,11 +5,11 @@ namespace Toggl.Multivac.Models
 {
     public interface ITimeEntry : IBaseModel
     {
-        int WorkspaceId { get; }
+        long WorkspaceId { get; }
 
-        int? ProjectId { get; }
+        long? ProjectId { get; }
 
-        int? TaskId { get; }
+        long? TaskId { get; }
 
         bool Billable { get; }
 
@@ -29,7 +29,7 @@ namespace Toggl.Multivac.Models
 
         DateTimeOffset? ServerDeletedAt { get; }
 
-        int UserId { get; }
+        long UserId { get; }
 
         string CreatedWith { get; }
     }
