+++ Toggl.PrimeRadiant/IRepository.cs
@@ -8,7 +8,7 @@ namespace Toggl.PrimeRadiant
     public interface IRepository<TModel>
         where TModel : IBaseModel, IDatabaseSyncable
     {
-        IObservable<TModel> GetById(int id);
+        IObservable<TModel> GetById(long id);
         IObservable<TModel> Create(TModel entity);
         IObservable<TModel> Update(TModel entity);
         IObservable<Unit> Delete(TModel entity);
