+++ Toggl.Ultrawave/Models/Client.cs
@@ -6,10 +6,10 @@ namespace Toggl.Ultrawave.Models
 {
     public sealed partial class Client : IClient
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("wid")]
-        public int WorkspaceId { get; set; }
+        public long WorkspaceId { get; set; }
 
         public string Name { get; set; }
