+++ Toggl.Ultrawave/Models/Tag.cs
@@ -6,9 +6,9 @@ namespace Toggl.Ultrawave.Models
 {
     public sealed partial class Tag : ITag
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int WorkspaceId { get; set; }
+        public long WorkspaceId { get; set; }
 
         public string Name { get; set; }
