+++ Toggl.Ultrawave/Models/Task.cs
@@ -6,18 +6,18 @@ namespace Toggl.Ultrawave.Models
 {
     public partial class Task : ITask
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
         [JsonProperty("pid")]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
 
         [JsonProperty("wid")]
-        public int WorkspaceId { get; set; }
+        public long WorkspaceId { get; set; }
 
         [JsonProperty("uid")]
-        public int? UserId { get; set; }
+        public long? UserId { get; set; }
 
         public int EstimatedSeconds { get; set; }
