+++ Toggl.Ultrawave/Models/User.cs
@@ -7,11 +7,11 @@ namespace Toggl.Ultrawave.Models
 {
     public sealed partial class User : IUser
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string ApiToken { get; set; }
         
-        public int DefaultWorkspaceId { get; set; }
+        public long DefaultWorkspaceId { get; set; }
 
         public string Email { get; set; }
